DAL layer update
BillDao, DishDao, DrinkDao, EmployeeDao, MenuDao, MenuItemDao, OrderDao, PaymentDao, TableDao added

+++ ChapeauModel/Bill.cs
@@ -3,9 +3,9 @@
     public class Bill
     {
         public int Id { get; set; }
-        public double TotalCost { get; set; }
-        public double SplitAmount { get; set; }
-        public double TipAmount { get; set; }
+        public float TotalCost { get; set; }
+        public float SplitAmount { get; set; }
+        public float TipAmount { get; set; }
 
         public override string ToString()
         {
